requirements.txt
mysql-connector-python
tkinter
os
pandas